Create Associative Array 10
Write a function that stores cars in garages. You will be given an array of strings. Each string will contain a number of a garage and info about a car. You have to store the car (with its info) in the given garage. The info about the car will be in the format:

"{key1}: {value1}, {key2}: {value2}…"

If the garage does not exist, create it. The cars will always be unique. At the end print the result in the format: "Garage № {number}: --- {carOneKeyOne} - {carOneValueOne}, {carOneKeyTwo} - {carOneValueTwo}… --- {the same for the next car} Garage № {number}: …"